* commit 'adc2e651aaa02bab98a3d342ec64f373c866a8c9': (123 commits)
  write comments
  es plugin migration
  move pluginDAO
  move pluginDAO
  es sample
  es migration as a sample
  Fix bugs in Template Registry
  refactor unit test
  BugFix in Template Registry JPA
  -m SUBUTAI-1676 pre-final commit for testing jpa in templateRegistry module
  -m SUBUTAI-1676 pre-final commit for testing jpa in templateRegistry module
  -m SUBUTAI-1676 pre-final commit for testing jpa in templateRegistry module
  Script Manager in progress
  -m SUBUTAI-1676 testing template registry with jpa
  add lock_test.sh
  change lock_test.sh directory
  delete test_master_destroy
  Script Manager in progress
  Added Script Manager
  Added Script Manager
  ...

+++ management/server/core/environment-manager/environment-manager-impl/src/main/java/org/safehaus/subutai/core/environment/impl/EnvironmentManagerImpl.java
@@ -21,10 +21,10 @@ import org.safehaus.subutai.common.protocol.EnvironmentBuildTask;
 import org.safehaus.subutai.common.protocol.NodeGroup;
 import org.safehaus.subutai.common.protocol.Template;
 import org.safehaus.subutai.common.tracker.TrackerOperation;
-import org.safehaus.subutai.core.container.api.ContainerCreateException;
 import org.safehaus.subutai.core.environment.api.EnvironmentManager;
 import org.safehaus.subutai.core.environment.api.TopologyEnum;
 import org.safehaus.subutai.core.environment.api.exception.EnvironmentBuildException;
+import org.safehaus.subutai.core.environment.api.exception.EnvironmentDestroyException;
 import org.safehaus.subutai.core.environment.api.exception.EnvironmentManagerException;
 import org.safehaus.subutai.core.environment.api.exception.EnvironmentPersistenceException;
 import org.safehaus.subutai.core.environment.api.helper.Environment;
@@ -32,12 +32,13 @@ import org.safehaus.subutai.core.environment.api.helper.EnvironmentBuildProcess;
 import org.safehaus.subutai.core.environment.api.helper.EnvironmentStatusEnum;
 import org.safehaus.subutai.core.environment.impl.builder.TopologyBuilder;
 import org.safehaus.subutai.core.environment.impl.dao.EnvironmentDAO;
-import org.safehaus.subutai.core.network.api.NetworkManager;
-import org.safehaus.subutai.core.network.api.NetworkManagerException;
 import org.safehaus.subutai.core.peer.api.ContainerHost;
 import org.safehaus.subutai.core.peer.api.Peer;
+import org.safehaus.subutai.core.peer.api.PeerException;
 import org.safehaus.subutai.core.peer.api.PeerManager;
 import org.safehaus.subutai.core.registry.api.TemplateRegistry;
+import org.safehaus.subutai.core.security.api.SecurityManager;
+import org.safehaus.subutai.core.security.api.SecurityManagerException;
 import org.safehaus.subutai.core.tracker.api.Tracker;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -64,7 +65,7 @@ public class EnvironmentManagerImpl implements EnvironmentManager
     //    private ServiceLocator serviceLocator;
     private EnvironmentDAO environmentDAO;
     private TemplateRegistry templateRegistry;
-    private NetworkManager networkManager;
+    private SecurityManager securityManager;
     private Tracker tracker;
     private DataSource dataSource;
 
@@ -88,15 +89,15 @@ public class EnvironmentManagerImpl implements EnvironmentManager
     }
 
 
-    public NetworkManager getNetworkManager()
+    public SecurityManager getSecurityManager()
     {
-        return networkManager;
+        return securityManager;
     }
 
 
-    public void setNetworkManager( final NetworkManager networkManager )
+    public void setSecurityManager( final SecurityManager securityManager )
     {
-        this.networkManager = networkManager;
+        this.securityManager = securityManager;
     }
 
 
@@ -116,12 +117,6 @@ public class EnvironmentManagerImpl implements EnvironmentManager
     {
         try
         {
-            //            this.serviceLocator = new ServiceLocator();
-            //            this.peerManager = serviceLocator.getService( PeerManager.class );
-            //            this.tracker = serviceLocator.getService( Tracker.class );
-            //            this.templateRegistry = serviceLocator.getService( TemplateRegistry.class );
-            //            this.networkManager = serviceLocator.getService( NetworkManager.class );
-
             this.topologyBuilder = new TopologyBuilder( this );
             this.environmentDAO = new EnvironmentDAO( dataSource );
         }
@@ -136,7 +131,7 @@ public class EnvironmentManagerImpl implements EnvironmentManager
     {
         this.environmentDAO = null;
         this.templateRegistry = null;
-        this.networkManager = null;
+        this.securityManager = null;
         this.peerManager = null;
         this.tracker = null;
         this.topologyBuilder = null;
@@ -187,29 +182,38 @@ public class EnvironmentManagerImpl implements EnvironmentManager
 
 
     @Override
-    public boolean destroyEnvironment( final UUID uuid )
+    public boolean destroyEnvironment( final UUID uuid ) throws EnvironmentDestroyException
     {
         Environment environment = getEnvironmentByUUID( uuid );
         int count = 0;
-        /*for ( EnvironmentContainer container : environment.getContainers() )
+        for ( ContainerHost container : environment.getContainers() )
         {
-            DestroyContainersMessage dcm =
-                    new DestroyContainersMessage( PeerCommandType.DESTROY, environment.getUuid(), container.getPeerId(),
-                            container.getAgentId() );
-            dcm.setHostname( container.getHostname() );
-            peerCommandDispatcher.invoke( dcm, 1000 * 60 );
-            if ( dcm.isSuccess() )
+            String ip = null;
+            try
             {
+                ip = container.getPeer().getPeerInfo().getIp();
+                container.dispose();
+                System.out.println( String.format( "Container %s destroyed.", container.getHostname() ) );
                 count++;
             }
-        }*/
+            catch ( PeerException e )
+            {
+                LOG.error( String.format( "Could not destroy container %s on %s: %s", container.getHostname(), ip,
+                        e.toString() ) );
+                throw new EnvironmentDestroyException( e.getMessage() );
+            }
+        }
 
-        //TODO: fix workaround
-        /*if ( count == environment.getContainers().size() )
+
+        if ( count == environment.getContainers().size() )
         {
-            return environmentDAO.deleteInfo( ENVIRONMENT, uuid );
-        }*/
-        return environmentDAO.deleteInfo( ENVIRONMENT, uuid.toString() );
+            return environmentDAO.deleteInfo( ENVIRONMENT, uuid.toString() );
+        }
+        else
+        {
+            throw new EnvironmentDestroyException( String.format( "Only %d out of %d containers destroyed.", count,
+                    environment.getContainers().size() ) );
+        }
     }
 
 
@@ -354,6 +358,7 @@ public class EnvironmentManagerImpl implements EnvironmentManager
                     {
                         for ( ContainerHost container : containers )
                         {
+                            container.setNodeGroupName( ccm.getNodeGroupName() );
                             environment.addContainer( container );
                         }
                     }
@@ -365,7 +370,7 @@ public class EnvironmentManagerImpl implements EnvironmentManager
                                 String.format( "FAILED creating environment on %s", ccm.getPeerId() ) );
                     }
                 }
-                catch ( ContainerCreateException e )
+                catch ( PeerException e )
                 {
                     LOG.error( e.getMessage(), e );
                     operation.addLogFailed( "Error occured while invoking command." );
@@ -387,9 +392,9 @@ public class EnvironmentManagerImpl implements EnvironmentManager
                 {
                     try
                     {
-                        networkManager.configSshOnAgents( environment.getContainers() );
+                        securityManager.configSshOnAgents( environment.getContainers() );
                     }
-                    catch ( NetworkManagerException e )
+                    catch ( SecurityManagerException e )
                     {
                         throw new EnvironmentBuildException( e.getMessage() );
                     }
@@ -398,9 +403,9 @@ public class EnvironmentManagerImpl implements EnvironmentManager
                 {
                     try
                     {
-                        networkManager.configHostsOnAgents( environment.getContainers(), blueprint.getDomainName() );
+                        securityManager.configHostsOnAgents( environment.getContainers(), blueprint.getDomainName() );
                     }
-                    catch ( NetworkManagerException e )
+                    catch ( SecurityManagerException e )
                     {
                         throw new EnvironmentBuildException( e.getMessage() );
                     }
